+++ src/planners/trajopt/trajopt_sco/src/optimizers.cpp
@@ -173,7 +173,7 @@ namespace sco
   }
   void Optimizer::callCallbacks(DblVec &x)
   {
-    for (int i = 0; i < callbacks_.size(); ++i)
+    for (size_t i = 0; i < callbacks_.size(); ++i)
     {
       callbacks_[i](prob_.get(), x);
     }
@@ -238,7 +238,7 @@ struct MultiCritFilter {
     double leastImprovement=INFINITY;
     BOOST_FOREACH(const DblVec& olderrvec, errvecs) {
       double improvement=0;
-      for (int i=0; i < errvec.size(); ++i) improvement += pospart(olderrvec[i] - errvec[i]);
+      for (size_t i=0; i < errvec.size(); ++i) improvement += pospart(olderrvec[i] - errvec[i]);
       leastImprovement = fmin(leastImprovement, improvement);
     }
     return leastImprovement;
@@ -351,7 +351,7 @@ struct MultiCritFilter {
           {
             DblVec cnt_costs1 = evaluateModelCosts(cnt_cost_models, model_var_vals);
             DblVec cnt_costs2 = model_cnt_viols;
-            for (int i = 0; i < cnt_costs2.size(); ++i)
+            for (size_t i = 0; i < cnt_costs2.size(); ++i)
               cnt_costs2[i] *= sqp_config.merit_error_coeff_;
             LOG_DEBUG("SHOULD BE ALMOST THE SAME: %s ?= %s", CSTR(cnt_costs1), CSTR(cnt_costs2));
             // not exactly the same because cnt_costs1 is based on aux variables, but they might not be at EXACTLY the right value
