+++ src/planners/trajopt/trajopt_utils/include/utils/eigen_slicing.hpp
@@ -7,7 +7,7 @@ namespace util
   VectorT fancySlice(const VectorT &x, const std::vector<int> &inds)
   {
     VectorT out(inds.size());
-    for (int i = 0; i < inds.size(); ++i)
+    for (size_t i = 0; i < inds.size(); ++i)
       out[i] = x[inds[i]];
     return out;
   }
@@ -16,7 +16,7 @@ namespace util
   std::vector<int> flatnonzero(const VectorT &x)
   {
     std::vector<int> out;
-    for (int i = 0; i < x.size(); ++i)
+    for (size_t i = 0; i < x.size(); ++i)
       if (x[i] != 0)
         out.push_back(i);
     return out;
